Delete README.md
Actor moves around a grid and encounters two percepts, Breeze and Stench, which indicate that a hazard, Pit or Wumpus respectively, are in an adjacent tile. The goal of the game is to gain the most points, which entails find the random tile with gold and return to the starting position at location (1,1). The game will end early and deduct points if the Actor steps into the tile with Pit or Wumpus. Only one Wumpus is present in each board and can be killed by shooting an arrow at it.